Refactor and enhance Noesis libc and quantum components
- Updated Makefile to dynamically include quantum field object files and link with nlibc_stubs library.
- Removed obsolete header files: backend.h and quantum.h from backups.
- Changed permissions of build_all.sh to make it executable.
- Cleaned up debug files by removing commented-out code.
- Added new Fish scripts for building nlibc stubs and fixing stdio types.
- Updated quantum_main.c to use NLIBC_FILE instead of FILE for file operations.
- Modified noesis_libc.h to prevent including system headers and added explicit declarations for memory management functions.
- Adjusted noesis_names.h to redefine getenv and added explicit declarations for memory management functions.
- Refactored stdio.h and stdio.c to replace FILE with NLIBC_FILE and updated related function signatures.
- Implemented minimal stubs for nlibc functions in nlibc_minimal_stubs.c.
- Enhanced assembly files for I/O operations, memory copying, and other utility functions with improved comments and structure.
- Updated quantum_field.c to ensure proper casting for memory allocation.
- Cleaned up various assembly files by removing unnecessary comments and ensuring consistency in documentation.

+++ source/quantum/field/quantum_field.c
@@ -8,6 +8,8 @@
 // Include stdlib.h which provides malloc and free declarations needed for new and del macros
 #include <stdlib/stdlib.h>
 #include <noesis_libc.h>  // Include all noesis_libc functionality with short names
+#include "../../../noesis_libc/include/noesis_names.h" // Ensure proper inclusion of nlibc_malloc and nlibc_free
+#include <stdint.h> // For uintptr_t
 
 // Helper function - allocate memory for a field
 static float* allocate_field_memory(int dimensions, int* size) {
@@ -15,12 +17,12 @@ static float* allocate_field_memory(int dimensions, int* size) {
     for (int i = 0; i < dimensions && i < 3; i++) {
         total_size *= size[i];
     }
-    return (float*)new(sizeof(float) * total_size);
+    return (float*)(uintptr_t)new(sizeof(float) * total_size); // Ensure proper casting
 }
 
 // Initialize a quantum field
 QuantumField* qf_create(int dimensions, int* size) {
-    QuantumField* field = (QuantumField*)new(sizeof(QuantumField));
+    QuantumField* field = (QuantumField*)(uintptr_t)new(sizeof(QuantumField)); // Ensure proper casting
     if (!field) return NULL;
 
     field->dimensions = dimensions > 3 ? 3 : dimensions;
